+++ hasoffer.admin/src/main/java/hasoffer/admin/worker/FixSkuErrorInPriceWorker.java
@@ -56,7 +56,7 @@ public class FixSkuErrorInPriceWorker implements Runnable {
                     continue;
                 }
 
-                float basePrice = -5.5f;
+                double basePrice = -5.5;
 
                 for (int i = skus.size() - 1; i >= 0; i--) {
