+++ hasoffer.core/src/main/java/hasoffer/core/product/iml/DataFixServiceImpl.java
@@ -185,8 +185,8 @@ public class DataFixServiceImpl implements IDataFixService {
             }
         });
 
-        float maxPrice = skuBoList.get(skuBoList.size() - 1).getPrice();
-        float minPrice = skuBoList.get(0).getPrice();
+        double maxPrice = skuBoList.get(skuBoList.size() - 1).getPrice();
+        double minPrice = skuBoList.get(0).getPrice();
 
         if (maxPrice / minPrice < 5) {
             return null;
