+++ hasoffer.core/src/main/java/hasoffer/core/search/SearchProductHelper.java
@@ -34,7 +34,7 @@ public class SearchProductHelper {
     public static Map<Website, ListProduct> getProducts(SrmSearchLog searchLog) {
 
         String keyword = searchLog.getKeyword();
-        float stdPrice = searchLog.getPrice();
+        double stdPrice = searchLog.getPrice();
 
         Map<Website, ListProduct> listProductMap = new HashMap<Website, ListProduct>();
 
@@ -53,7 +53,7 @@ public class SearchProductHelper {
         return listProductMap;
     }
 
-    public static void getProductsFromAffiliate(Map<Website, ListProduct> listProductMap, String keyword, float stdPrice) {
+    public static void getProductsFromAffiliate(Map<Website, ListProduct> listProductMap, String keyword, double stdPrice) {
         try {
             getProductFromFlipkart(listProductMap, keyword, stdPrice);
         } catch (Exception e) {
@@ -61,7 +61,7 @@ public class SearchProductHelper {
         }
     }
 
-    public static void getProductsFromMSP(Map<Website, ListProduct> listProductMap, String keyword, float stdPrice) {
+    public static void getProductsFromMSP(Map<Website, ListProduct> listProductMap, String keyword, double stdPrice) {
         MspListProcessor listProcessor = new MspListProcessor();
         try {
             List<ListProduct> listProducts = listProcessor.getProductSetByKeyword(keyword, 5);
@@ -102,7 +102,7 @@ public class SearchProductHelper {
         }
     }
 
-    public static void getProductsFromWebsite(Map<Website, ListProduct> listProductMap, String keyword, float stdPrice) {
+    public static void getProductsFromWebsite(Map<Website, ListProduct> listProductMap, String keyword, double stdPrice) {
         //遍历websiteList，添加比较列表
         for (Website website : websites) {
 
@@ -134,7 +134,7 @@ public class SearchProductHelper {
         }
     }
 
-    public static void getProductFromFlipkart(Map<Website, ListProduct> listProductMap, String keyword, float stdPrice) throws Exception {
+    public static void getProductFromFlipkart(Map<Website, ListProduct> listProductMap, String keyword, double stdPrice) throws Exception {
         Website website = Website.FLIPKART;
 
         FlipkartAffiliateProductProcessor affProcessor = new FlipkartAffiliateProductProcessor();
@@ -158,7 +158,7 @@ public class SearchProductHelper {
         }
     }
 
-    private static boolean nearPrice(float stdPrice, float price) {
+    private static boolean nearPrice(double stdPrice, float price) {
 
         if (stdPrice == 0) {
             return true;
